update project structure and makefile

+++ srcs/spec_c.c
@@ -6,7 +6,7 @@
 /*   By: rschuppe <rschuppe@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/23 17:09:14 by rschuppe          #+#    #+#             */
-/*   Updated: 2018/12/27 20:17:03 by rschuppe         ###   ########.fr       */
+/*   Updated: 2018/12/27 21:42:58 by rschuppe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	print_char(void *value, t_spec *spec, int *len)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (CHECK_FLAG(FLAG_MINUS))
